[JAX Runner] Refactor jax runner to delegate precompilation to compilation manager.
